ref: refactor done and tests passed
