+++ src/tests.rs
@@ -11,7 +11,7 @@ use chrono::*;
 #[cfg(test)]
 #[test]
 fn get_next_monday() {
-    use crate::models::slot_iterator::get_start_of_repeat_step;
+    use crate::models::slot::utils::get_start_of_repeat_step;
 
     let monday = NaiveDate::from_ymd_opt(2022, 09, 26)
         .unwrap()
@@ -47,7 +47,7 @@ fn get_next_monday() {
 
 #[test]
 fn get_next_weekend() {
-    use crate::models::slot_iterator::get_start_of_repeat_step;
+    use crate::models::slot::utils::get_start_of_repeat_step;
 
     let monday = NaiveDate::from_ymd_opt(2022, 09, 26)
         .unwrap()
